Split utils across DCCs to align with Unreal
