+++ doc/developing/api_references/ftrack_connect_pipeline_houdini.utils.rst
@@ -5,3 +5,13 @@ ftrack\_connect\_pipeline\_houdini.utils package
    :members:
    :undoc-members:
    :show-inheritance:
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 1
+
+   ftrack_connect_pipeline_houdini.plugin.bootstrap
+   ftrack_connect_pipeline_houdini.plugin.file
+   ftrack_connect_pipeline_houdini.plugin.node
