+++ source/ftrack_connect_pipeline_houdini/constants/asset/modes.py
@@ -1,9 +1,7 @@
 # :coding: utf-8
 # :copyright: Copyright (c) 2014-2021 ftrack
 
-from ftrack_connect_pipeline_houdini.utils import (
-    custom_commands as houdini_utils,
-)
+from ftrack_connect_pipeline_houdini import utils as houdini_utils
 
 # Load Modes
 MERGE_MODE = 'merge'
