+++ source/ftrack_connect_pipeline_houdini/plugin/__init__.py
@@ -6,9 +6,7 @@ from ftrack_connect_pipeline_qt import plugin as pluginWidget
 from ftrack_connect_pipeline_houdini import constants as houdini_constants
 from ftrack_connect_pipeline_houdini.asset import HoudiniFtrackObjectManager
 from ftrack_connect_pipeline_houdini.asset.dcc_object import HoudiniDccObject
-from ftrack_connect_pipeline_houdini.utils import (
-    custom_commands as houdini_utils,
-)
+from ftrack_connect_pipeline_houdini import utils as houdini_utils
 
 
 class HoudiniBasePlugin(plugin.BasePlugin):
