+++ source/ftrack_connect_pipeline_houdini/plugin/open/importer.py
@@ -8,11 +8,8 @@ from ftrack_connect_pipeline_houdini.plugin import (
     HoudiniBasePluginWidget,
 )
 
-from ftrack_connect_pipeline_houdini.utils import (
-    custom_commands as houdini_utils,
-)
+from ftrack_connect_pipeline_houdini import utils as houdini_utils
 from ftrack_connect_pipeline_houdini.constants.asset import modes as load_const
-from ftrack_connect_pipeline_houdini.constants import asset as asset_const
 
 
 class HoudiniOpenerImporterPlugin(
